feat: update launch settings and package versions
Disable the launch browser in launchSettings.json and change the 
launch URL to "scalar" for improved API access. Update package 
versions in Directory.Packages.props to ensure compatibility 
with the latest dependencies, including adding Scalar.AspNetCore 
for enhanced functionality. Remove unused WebAppExtensions 
class to streamline the codebase.
