+++ Directory.Packages.props
@@ -12,7 +12,6 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
   <ItemGroup>
-    
     <PackageVersion Include="Blazored.SessionStorage" Version="2.4.0" />
     <PackageVersion Include="Bogus" Version="35.6.1" />
     <PackageVersion Include="Brushtail.FontAwesome.Blazor" Version="1.0.19" />
@@ -31,6 +30,7 @@
     <PackageVersion Include="Microsoft.Extensions.Http" Version="9.0.0" />
     <PackageVersion Include="MySqlConnector" Version="2.4.0" />
     <PackageVersion Include="Polly" Version="8.5.0" />
+    <PackageVersion Include="Scalar.AspNetCore" Version="2.0.9" />
     <PackageVersion Include="Serilog" Version="4.2.0" />
     <PackageVersion Include="Serilog.AspNetCore" Version="9.0.0" />
     <PackageVersion Include="Swashbuckle.AspNetCore" Version="7.2.0" />
@@ -50,9 +50,9 @@
     <PackageVersion Include="Microsoft.EntityFrameworkCore.Sqlite" Version="$(EfVersion)" />
     <PackageVersion Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="$(EfVersion)" />
     <!-- Version together with Aspire -->
-<!--    <PackageVersion Include="Aspire.Hosting" Version="$(AspireVersion)" />-->
+    <!--    <PackageVersion Include="Aspire.Hosting" Version="$(AspireVersion)" />-->
     <PackageVersion Include="Aspire.Hosting.Testing" Version="$(AspireVersion)" />
-<!--    <PackageVersion Include="Aspire.AppHost.Sdk" Version="$(AspireVersion)" />-->
+    <!--    <PackageVersion Include="Aspire.AppHost.Sdk" Version="$(AspireVersion)" />-->
     <PackageVersion Include="Aspire.Hosting.AppHost" Version="$(AspireVersion)" />
     <PackageVersion Include="Aspire.Hosting.MySql" Version="$(AspireVersion)" />
     <PackageVersion Include="Aspire.Hosting.NodeJs" Version="$(AspireVersion)" />
@@ -84,7 +84,7 @@
     <PackageVersion Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="$(AspnetVersion)" />
     <PackageVersion Include="Microsoft.AspNetCore.Identity.UI" Version="$(AspnetVersion)" />
     <PackageVersion Include="Microsoft.AspNetCore.Mvc.Testing" Version="$(AspnetVersion)" />
-    <PackageVersion Include="Microsoft.AspNetCore.OpenApi" Version="$(AspnetVersion)" />
+    <PackageVersion Include="Microsoft.AspNetCore.OpenApi" Version="9.0.1" />
     <PackageVersion Include="Microsoft.AspNetCore.TestHost" Version="$(AspnetVersion)" />
     <PackageVersion Include="Microsoft.Extensions.Http.Resilience" Version="$(AspnetVersion)" />
     <PackageVersion Include="Microsoft.Extensions.Identity.Stores" Version="$(AspnetVersion)" />
