+++ Source/Core/Core.csproj
@@ -22,6 +22,7 @@
         <PackageReference Include="OpenTelemetry.Instrumentation.Http"/>
         <PackageReference Include="OpenTelemetry.Instrumentation.Process"/>
         <PackageReference Include="OpenTelemetry.Instrumentation.Runtime"/>
+        <PackageReference Include="Scalar.AspNetCore" />
         <PackageReference Include="Swashbuckle.AspNetCore"/>
         <PackageReference Include="Swashbuckle.AspNetCore.Filters" />
         <PackageReference Include="Swashbuckle.AspNetCore.SwaggerGen"/>
